Add files via upload
Here is a Simplify 3D profile I use with Maker Geeks Raptor PLA 1.75.  The increased PLA temperature, decreased bed temperature, and decreased fan speeds are working well for me.  First layer (90%) is looking really good with this profile but I have seen good results with a higher percentage also.  You might tweak that setting to your liking.  I feel this success is dependent on a correct/close to correct Live Z setting on the MK3.

Since each filament manufacturer is different, this may not work as is with other brands.